# openapi.yaml
openapi: 3.0.3
info:
  title: API ИИ-гида музея Фаберже
  version: 0.1.0-mvp
  description: |
    Backend-API мобильного web-приложения (PWA) «ИИ-гид музея Фаберже».

    Посетитель сканирует **QR-код** на входе экспозиции, попадает в приложение
    без регистрации и установки, перемещается по **интерактивной карте**
    (зал → витрина → экспонат), а главное — может **сфотографировать экспонат**,
    получить о нём рассказ от ИИ-гида и **прослушать** его озвучку.

    Этот документ описывает контракт backend-сервиса (FastAPI на Yandex Cloud),
    к которому обращается фронтенд. Спецификация написана по дорожной карте MVP
    (25 мая – 25 июня) и служит единым контрактом между Frontend, Backend и ML.

    ## Ключевой пользовательский сценарий (бесшовное распознавание)

    ```
    Клик «Распознать экспонат»
      → камера (WebRTC)
      → снимок
      → POST /recognition (фото → YOLO → label_slug)
      → POST /guide/story (label_slug → raw_history из БД → YandexGPT → рассказ)
      → показ рассказа + 3–4 всплывающих вопроса
      → POST /speech («Прослушать», SpeechKit → аудио)
      → блок «Другие экспонаты зала» (GET /exhibits/{id}/related)
    ```

    ## Доменная модель

    Иерархия каталога повторяет физическую структуру музея:

    | Уровень   | Сущность   | Таблица БД   | Ключевые поля                                   |
    |-----------|------------|--------------|-------------------------------------------------|
...

...
    (через публичный API не отдаётся).

...

...

...

...

...

...

...

...

...

...

...
servers:
...
security: []
paths:
...
  /halls:
    get:
      tags: [Карта и навигация]
      summary: Список залов
      description: Постраничный список залов музея для карты и навигации.
      operationId: listHalls
      security: []
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Список залов.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HallListResponse'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
...
  /exhibits:
    get:
      tags: [Экспонаты]
      summary: Список / фильтрация экспонатов
      description: >-
        Постраничный список экспонатов с фильтрами по залу, витрине,
        `label_slug` и подстроке названия. Универсальная выборка для каталога.
      operationId: listExhibits
      security: []
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - name: hall_id
          in: query
          description: Фильтр по идентификатору зала.
          required: false
          schema:
            type: integer
            minimum: 1
          example: 5
        - name: showcase_id
          in: query
          description: Фильтр по идентификатору витрины.
          required: false
          schema:
            type: integer
            minimum: 1
        - name: label_slug
          in: query
          description: Фильтр по точному `label_slug` (то, что возвращает YOLO).
          required: false
          schema:
            type: string
          example: faberge_egg_coronation
        - name: q
          in: query
          description: Поиск по подстроке названия экспоната.
          required: false
          schema:
            type: string
          example: яйцо
      responses:
        '200':
          description: Список экспонатов.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExhibitListResponse'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /exhibits/{exhibit_id}:
    get:
      tags: [Экспонаты]
      summary: Карточка экспоната
      description: >-
        Полные данные карточки: фото, 3D-модель (Koinovo), краткое описание
        (год, мастер, материалы), ссылки на зал и витрину. Поле `raw_history`
        в публичный ответ не включается.
      operationId: getExhibit
      security: []
      parameters:
        - $ref: '#/components/parameters/ExhibitId'
      responses:
        '200':
          description: Карточка экспоната.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exhibit'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /exhibits/by-slug/{label_slug}:
    get:
      tags: [Экспонаты]
      summary: Экспонат по label_slug
      description: >-
        Получение карточки по `label_slug` — удобно сразу после распознавания,
        когда YOLO вернул класс, но `exhibit_id` ещё неизвестен фронтенду.
      operationId: getExhibitBySlug
      security: []
      parameters:
        - $ref: '#/components/parameters/LabelSlug'
      responses:
        '200':
          description: Карточка экспоната.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exhibit'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /exhibits/{exhibit_id}/related:
    get:
      tags: [Экспонаты]
      summary: Другие экспонаты зала
      description: >-
        Соседние экспонаты того же зала (по `hall_id`), исключая текущий.
        Питает блок «Другие экспонаты зала» внизу карточки — карусель/сетку
        кликабельных фото для удержания посетителя в каталоге.
      operationId: listRelatedExhibits
      security: []
      parameters:
        - $ref: '#/components/parameters/ExhibitId'
        - name: limit
          in: query
          description: Сколько соседних экспонатов вернуть.
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 12
      responses:
        '200':
          description: Соседние экспонаты зала.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExhibitListResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'

  # ───────────────────────────── Поиск ───────────────────────────────────
...
  /admin/halls:
    post:
      tags: [Администрирование]
      summary: '[Вне MVP] Создать зал'
      operationId: createHall
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HallCreate'
      responses:
        '201':
          description: Зал создан.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HallDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /admin/showcases:
    post:
      tags: [Администрирование]
      summary: '[Вне MVP] Создать витрину'
      operationId: createShowcase
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ShowcaseCreate'
      responses:
        '201':
          description: Витрина создана.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShowcaseDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /admin/exhibits:
    post:
      tags: [Администрирование]
      summary: '[Вне MVP] Создать экспонат'
      description: >-
        Создание экспоната, включая внутреннее поле `raw_history`
        (факты для генерации рассказа LLM).
      operationId: createExhibit
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExhibitCreate'
            examples:
              winterEgg:
                summary: Зимнее яйцо с raw_history
                value:
                  showcase_id: 12
                  label_slug: faberge_egg_winter
                  name: Яйцо «Зимнее»
                  year_created: 1913
                  master_name: Альма Пиль
                  material: Горный хрусталь, платина, бриллианты
                  short_description: >-
                    Пасхальное яйцо в виде ледяного кристалла с сюрпризом-корзиной
                    подснежников.
                  raw_history: >-
                    Заказчик — Николай II. Подарок императрице Марии Фёдоровне.
                    Дизайн — Альма Пиль. Использованы горный хрусталь и бриллианты,
                    имитирующие иней. Сюрприз — корзина подснежников.
      responses:
        '201':
          description: Экспонат создан.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExhibitAdmin'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /admin/exhibits/{exhibit_id}:
    put:
      tags: [Администрирование]
      summary: '[Вне MVP] Полностью обновить экспонат'
      operationId: updateExhibit
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ExhibitId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExhibitUpdate'
      responses:
        '200':
          description: Экспонат обновлён.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExhibitAdmin'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
    patch:
      tags: [Администрирование]
      summary: '[Вне MVP] Частично обновить экспонат'
      operationId: patchExhibit
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ExhibitId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExhibitPatch'
      responses:
        '200':
          description: Экспонат обновлён.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExhibitAdmin'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
    delete:
      tags: [Администрирование]
      summary: '[Вне MVP] Удалить экспонат'
      operationId: deleteExhibit
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ExhibitId'
      responses:
        '204':
          description: Экспонат удалён.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /admin/exhibits/{exhibit_id}/media:
    post:
      tags: [Администрирование]
      summary: '[Вне MVP] Загрузить фото экспоната'
      description: >-
        Загрузка изображения экспоната в Yandex Object Storage; возвращает
        публичные URL (оригинал + превью), раздаваемые через CDN.
      operationId: uploadExhibitMedia
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ExhibitId'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: Изображение (JPEG/PNG/WebP).
                is_primary:
                  type: boolean
                  default: false
                  description: Сделать это изображение главным в карточке.
      responses:
        '201':
          description: Файл загружен.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MediaUploadResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
  /admin/analytics/overview:
    get:
      tags: [Администрирование]
      summary: '[Вне MVP] Сводная аналитика'
      description: >-
        Агрегированные показатели использования за период: сессии,
        распознавания, доля успешных распознаваний, сообщения в чате,
        прослушивания аудио, топ экспонатов и залов.
      operationId: getAnalyticsOverview
      security:
        - bearerAuth: []
      parameters:
        - name: from
          in: query
          description: Начало периода (ISO-8601 дата).
          required: false
          schema:
            type: string
            format: date
          example: '2026-05-25'
        - name: to
          in: query
          description: Конец периода (ISO-8601 дата).
          required: false
          schema:
            type: string
            format: date
          example: '2026-06-25'
      responses:
        '200':
          description: Сводные показатели.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalyticsOverview'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'

  # ───────────────────────── Телеметрия [вне MVP] ────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Bearer-токен администратора. Только для эндпоинтов `/admin/**`.
  parameters:
    Limit:
      name: limit
      in: query
      description: Максимум элементов на странице.
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    Offset:
      name: offset
      in: query
      description: Смещение от начала выборки (постраничность).
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
...
    ExhibitId:
      name: exhibit_id
      in: path
      required: true
      description: Идентификатор экспоната.
      schema:
        type: integer
        minimum: 1
      example: 101
    LabelSlug:
      name: label_slug
      in: path
      required: true
      description: Стабильный slug класса от YOLO.
      schema:
        type: string
        pattern: '^[a-z0-9_]+$'
      example: faberge_egg_coronation
  responses:
    NotFound:
      description: Ресурс не найден.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: Экспонат не найден.
...
    Unauthorized:
      description: Требуется авторизация администратора.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: Not authenticated
    Conflict:
      description: Конфликт уникальности (например, дубль `label_slug` или номера).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: Экспонат с таким label_slug уже существует.
    PayloadTooLarge:
      description: Файл превышает допустимый размер (10 МБ).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: Размер файла превышает 10 МБ.
    UnsupportedMediaType:
      description: Неподдерживаемый формат файла.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: Поддерживаются только JPEG, PNG и WebP.
    UnprocessableEntity:
      description: Ошибка валидации запроса (формат FastAPI).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/HTTPValidationError'
...
  schemas:
...
    Hall:
      type: object
      required: [id, hall_number]
      properties:
        id:
          type: integer
          example: 5
        hall_number:
          type: integer
          description: Номер зала на плане экспозиции (уникальный).
          example: 5
        name:
          type: string
          nullable: true
          example: Синяя гостиная
        description:
          type: string
          nullable: true
          description: >-
            Описание зала для уровня зала на карте и для контекста ИИ-гида.
            Наполняется из плана экспозиции и путеводителя.
          example: >-
            Парадная гостиная Шуваловского дворца, где представлены девять
            императорских пасхальных яиц Фаберже.
        level:
          type: integer
          nullable: true
          description: Этаж / уровень в здании музея.
          example: 2
        cover_image_url:
          type: string
          format: uri
          nullable: true
          example: https://cdn.example.cloud/halls/blue/cover.jpg
        showcase_count:
          type: integer
          nullable: true
          example: 6
        exhibit_count:
          type: integer
          nullable: true
          example: 24
...
    HallDetail:
      allOf:
        - $ref: '#/components/schemas/Hall'
        - type: object
          properties:
            showcases:
              type: array
              description: Витрины зала (могут включаться по необходимости).
              items:
                $ref: '#/components/schemas/Showcase'
...
    HallListResponse:
      type: object
      required: [items, total, limit, offset]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Hall'
        total:
          type: integer
          example: 12
        limit:
          type: integer
          example: 20
        offset:
          type: integer
          example: 0

    # ── Витрины ──
...
    ShowcaseDetail:
      allOf:
        - $ref: '#/components/schemas/Showcase'
        - type: object
          properties:
            hall:
              $ref: '#/components/schemas/HallBrief'
            exhibits:
              type: array
              items:
                $ref: '#/components/schemas/ExhibitSummary'
...
    Image:
      type: object
      required: [url]
      properties:
        url:
          type: string
          format: uri
          example: https://cdn.example.cloud/exhibits/coronation/main.jpg
        alt:
          type: string
          nullable: true
          example: Коронационное яйцо, общий вид
        width:
          type: integer
          nullable: true
          example: 1600
        height:
          type: integer
          nullable: true
          example: 1600
...
    Exhibit:
      type: object
      description: >-
        Полная карточка экспоната для публичного API. Внутреннее поле
        `raw_history` сюда не включается.
      required: [id, name]
      properties:
        id:
          type: integer
          example: 101
        label_slug:
          type: string
          nullable: true
          description: Класс, который возвращает YOLO.
          example: faberge_egg_coronation
        name:
          type: string
          example: Яйцо «Коронационное»
        year_created:
          type: integer
          nullable: true
          example: 1897
        master_name:
          type: string
          nullable: true
          example: Михаил Перхин
        material:
          type: string
          nullable: true
          example: Золото, гильоше-эмаль, бриллианты, рубины
        short_description:
          type: string
          nullable: true
          example: >-
            Пасхальное яйцо-сюрприз, подаренное Николаем II императрице
            Александре Фёдоровне в память о коронации 1896 года.
        image_url:
          type: string
          format: uri
          nullable: true
          description: Главное изображение (Object Storage + CDN).
          example: https://cdn.example.cloud/exhibits/coronation/main.jpg
        images:
          type: array
          description: Галерея изображений.
          items:
            $ref: '#/components/schemas/Image'
        model_3d_url:
          type: string
          format: uri
          nullable: true
          description: Ссылка на 3D-модель Koinovo.
          example: https://koinovo.ru/fabergemuseum/coronation
        model_3d_embed:
          type: string
          nullable: true
          description: HTML/iframe embed-код 3D-модели (когда предоставит Koinovo).
        audio_url:
          type: string
          format: uri
          nullable: true
          description: Предсинтезированная озвучка краткого описания (если есть).
        source_url:
          type: string
          format: uri
          nullable: true
          description: Первоисточник описания на сайте музея.
          example: https://fabergemuseum.ru/kollekczii/shedevryi-kollekczii/
        hall:
          $ref: '#/components/schemas/HallBrief'
        showcase:
          $ref: '#/components/schemas/ShowcaseBrief'
    ExhibitListResponse:
      type: object
      required: [items, total, limit, offset]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ExhibitSummary'
        total:
          type: integer
          example: 24
        limit:
          type: integer
          example: 20
        offset:
          type: integer
          example: 0

    # ── Поиск ──
...
    HallCreate:
      type: object
      required: [hall_number]
      properties:
        hall_number:
          type: integer
          example: 5
        name:
          type: string
          nullable: true
          example: Синяя гостиная
        description:
          type: string
          nullable: true
        level:
          type: integer
          nullable: true
          example: 2
    ShowcaseCreate:
      type: object
      required: [hall_id, showcase_number]
      properties:
        hall_id:
          type: integer
          example: 5
        showcase_number:
          type: integer
          example: 3
        name:
          type: string
          nullable: true
    ExhibitCreate:
      type: object
      required: [showcase_id, name]
      properties:
        showcase_id:
          type: integer
          example: 12
        label_slug:
          type: string
          nullable: true
          pattern: '^[a-z0-9_]+$'
          example: faberge_egg_winter
        name:
          type: string
          example: Яйцо «Зимнее»
        year_created:
          type: integer
          nullable: true
          example: 1913
        master_name:
          type: string
          nullable: true
          example: Альма Пиль
        material:
          type: string
          nullable: true
          example: Горный хрусталь, платина, бриллианты
        short_description:
          type: string
          nullable: true
        raw_history:
          type: string
          nullable: true
          description: >-
            Внутренние факты для генерации рассказа LLM. Рекомендуется хранить
            буллитами или в JSON. Через публичный API не отдаётся.
          example: >-
            Заказчик — Николай II. Подарок императрице Марии Фёдоровне.
            Дизайн — Альма Пиль. Горный хрусталь и бриллианты. Сюрприз —
            корзина подснежников.
        image_url:
          type: string
          format: uri
          nullable: true
        model_3d_url:
          type: string
          format: uri
          nullable: true
    ExhibitUpdate:
      allOf:
        - $ref: '#/components/schemas/ExhibitCreate'
      description: Полная замена полей экспоната (PUT).
    ExhibitPatch:
      type: object
      description: Частичное обновление — все поля необязательны.
      properties:
        showcase_id:
          type: integer
        label_slug:
          type: string
          pattern: '^[a-z0-9_]+$'
        name:
          type: string
        year_created:
          type: integer
          nullable: true
        master_name:
          type: string
          nullable: true
        material:
          type: string
          nullable: true
        short_description:
          type: string
          nullable: true
        raw_history:
          type: string
          nullable: true
        image_url:
          type: string
          format: uri
          nullable: true
        model_3d_url:
          type: string
          format: uri
          nullable: true
    ExhibitAdmin:
      allOf:
        - $ref: '#/components/schemas/Exhibit'
        - type: object
          description: Админское представление — включает внутреннее `raw_history`.
          properties:
            raw_history:
              type: string
              nullable: true
              example: >-
                Заказчик — Николай II. Подарок императрице Марии Фёдоровне.
    MediaUploadResponse:
      type: object
      required: [image_url, object_key]
      properties:
        image_url:
          type: string
          format: uri
          example: https://cdn.example.cloud/exhibits/winter/main.jpg
        thumbnail_url:
          type: string
          format: uri
          nullable: true
          example: https://cdn.example.cloud/exhibits/winter/thumb.jpg
        object_key:
          type: string
          description: Ключ объекта в Yandex Object Storage.
          example: exhibits/winter/main.jpg
...
    AnalyticsOverview:
      type: object
      properties:
        from:
          type: string
          format: date
          nullable: true
          example: '2026-05-25'
        to:
          type: string
          format: date
          nullable: true
          example: '2026-06-25'
        total_sessions:
          type: integer
          example: 1240
        total_recognitions:
          type: integer
          example: 980
        recognition_success_rate:
          type: number
          format: float
          minimum: 0
          maximum: 1
          example: 0.86
        total_chat_messages:
          type: integer
          example: 3450
        total_audio_plays:
          type: integer
          example: 1120
        top_exhibits:
          type: array
          items:
            $ref: '#/components/schemas/AnalyticsTopItem'
        top_halls:
          type: array
          items:
            $ref: '#/components/schemas/AnalyticsTopItem'

    # ── Телеметрия ──
...
    HTTPValidationError:
      type: object
      description: Ошибка валидации запроса в формате FastAPI.
      properties:
        detail:
          type: array
          items:
            $ref: '#/components/schemas/ValidationErrorItem'